backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
firebase-admin==6.5.0
google-cloud-firestore==2.19.0
transformers==4.44.2
torch>=2.2.0
openai==1.51.0
google-generativeai==0.7.2
requests==2.32.3
pydantic==2.9.2
